LayoutTests/http/tests/w3c/html/dom/documents/dom-tree-accessors/document.getElementsByName/document.getElementsByName-namespace-expected.txt

a
b



FAIL getElementsByName and foreign namespaces assert_equals: expected 1 but got 2
